+++ visualization/app/codeCharta/codeCharta.service.ts
@@ -5,7 +5,6 @@ import { FileStateService } from "./state/fileState.service"
 export class CodeChartaService {
 	public static ROOT_NAME = "root"
 	public static ROOT_PATH = "/" + CodeChartaService.ROOT_NAME
-	public static SELECTOR = "codeChartaService"
 
 	constructor(private fileStateService: FileStateService) {}
 
@@ -36,7 +35,7 @@ export class CodeChartaService {
 			settings: {
 				fileSettings: {
 					edges: fileContent.edges || [],
-					attributeTypes: fileContent.attributeTypes || {},
+					attributeTypes: fileContent.attributeTypes || { nodes: [], edges: [] },
 					blacklist: fileContent.blacklist || [],
 					markedPackages: []
 				}