+++ visualization/app/codeCharta/state/metric.service.ts
@@ -1,10 +1,20 @@
-import { BlacklistItem, BlacklistType, CodeMapNode, FileState, MetricData, RecursivePartial, Settings } from "../codeCharta.model"
+import {
+	AttributeType,
+	BlacklistItem,
+	BlacklistType,
+	CodeMapNode,
+	FileState,
+	MetricData,
+	RecursivePartial,
+	Settings
+} from "../codeCharta.model"
 import { hierarchy, HierarchyNode } from "d3"
 import { FileStateService, FileStateServiceSubscriber } from "./fileState.service"
 import { FileStateHelper } from "../util/fileStateHelper"
 import { IAngularEvent, IRootScopeService } from "angular"
 import { SettingsService, SettingsServiceSubscriber } from "./settings.service"
 import { CodeMapHelper } from "../util/codeMapHelper"
+import _ from "lodash"
 
 export interface MetricServiceSubscriber {
 	onMetricDataAdded(metricData: MetricData[], event: IAngularEvent)
@@ -20,6 +30,7 @@ export class MetricService implements FileStateServiceSubscriber, SettingsServic
 	private static METRIC_DATA_ADDED_EVENT = "metric-data-added"
 	private static METRIC_DATA_REMOVED_EVENT = "metric-data-removed"
 
+	//TODO MetricData should contain attributeType
 	private metricData: MetricData[] = []
 
 	constructor(private $rootScope: IRootScopeService, private fileStateService: FileStateService) {
@@ -62,6 +73,15 @@ export class MetricService implements FileStateServiceSubscriber, SettingsServic
 	public getMaxMetricByMetricName(metricName: string): number {
 		const metric: MetricData = this.metricData.find(x => x.name == metricName)
 		return metric ? metric.maxValue : undefined
+	}
+
+	public getAttributeTypeByMetric(metricName: string, settings: Settings): AttributeType {
+		const attributeType = settings.fileSettings.attributeTypes.nodes.find(x => _.findKey(x) === metricName)
+
+		if (attributeType) {
+			return attributeType[metricName]
+		}
+		return null
 	}
 
 	private calculateMetrics(fileStates: FileState[], visibleFileStates: FileState[], blacklist: BlacklistItem[]): MetricData[] {